requirements.txt
streamlit
requests
PyJWT
Pillow
plotly
